Lots of cleanup, moved to deckpicker menu, various refactoring, various fixes
