+++ AnkiDroid/src/main/java/com/ichi2/libanki/Models.java
@@ -509,8 +509,9 @@ public class Models {
     }
 
 
-    public void setSortIdx(JSONObject m, int idx) {
+    public void setSortIdx(JSONObject m, int idx) throws ConfirmModSchemaException{
     try {
+        mCol.modSchema();
         m.put("sortf", idx);
         mCol.updateFieldCache(Utils.toPrimitive(nids(m)));
         save(m);
@@ -771,7 +772,7 @@ public class Models {
      * Removing a template
      *
      * @return False if removing template would leave orphan notes.
-     * @throws ConfirmModSchemaException 
+     * @throws ConfirmModSchemaException
      */
     public boolean remTemplate(JSONObject m, JSONObject template) throws ConfirmModSchemaException {
         try {
@@ -892,7 +893,7 @@ public class Models {
      * @param newModel For replacing the old model with another one. Should be self if the model is not changing
      * @param fmap Map for switching fields. This is ord->ord and there should not be duplicate targets
      * @param cmap Map for switching cards. This is ord->ord and there should not be duplicate targets
-     * @throws ConfirmModSchemaException 
+     * @throws ConfirmModSchemaException
      */
     public void change(JSONObject m, long[] nids, JSONObject newModel, Map<Integer, Integer> fmap, Map<Integer, Integer> cmap) throws ConfirmModSchemaException {
         mCol.modSchema(true);
