+++ AnkiDroid/src/main/res/menu/deck_picker.xml
@@ -45,6 +45,10 @@
         android:menuCategory="secondary"
         android:title="@string/backup_restore"/>
     <item
+        android:id="@+id/action_model_browser_open"
+        android:menuCategory="secondary"
+        android:title="@string/model_browser_label"/>
+    <item
         android:id="@+id/action_import"
         android:menuCategory="secondary"
         android:title="@string/menu_import"/>
