+++ AnkiDroid/src/main/res/values/01-core.xml
@@ -25,7 +25,6 @@
     <string name="settings">Settings</string>
     <string name="help">Help</string>
     <string name="send_feedback">Send feedback</string>
-    <string name="model_browser">Note type browser</string>
 
     <string name="studyoptions_title">Study options</string>
     <string name="studyoptions_due_today">Due today:</string>
