+++ AnkiDroid/src/main/res/values/17-model-manager.xml
@@ -2,8 +2,9 @@
 <resources>
 
     <!--Menu Labels-->
-    <string name="model_browser_label">Note Type Browser</string>
-    <string name="model_editor_label">Note Type Editor</string>
+    <string name="model_browser_label">Manage Note Types</string>
+    <string name="model_editor_label">Manage Fields</string>
+
 
     <!--Toasts-->
     <string name="toast_last_model">You cannot delete your last model</string>
@@ -13,18 +14,22 @@
     <string name="toast_out_of_range">You must enter a valid value</string>
 
 
+    <!--Plurals-->
+
     <plurals name="model_browser_types_available">
-        <item quantity="one">%d note types available</item>
+        <item quantity="one">%d note type available</item>
         <item quantity="other">%d note types available</item>
     </plurals>
 
     <plurals name="model_browser_of_type">
-        <item quantity="one">%d notes</item>
+        <item quantity="one">%d note</item>
         <item quantity="other">%d notes</item>
     </plurals>
 
+
     <!--Button Values-->
     <string name="model_editor_action_add_model">Add Note Type</string>
+
 
     <!--Browser-->
     <string name="model_browser_delete">Delete note type</string>
@@ -33,26 +38,31 @@
     <string name="model_browser_template">Edit cards</string>
     <string name="model_browser_add">Add note type</string>
 
+
     <!--Rename-->
     <string name="rename_model">Rename model</string>
 
-    <!--Field Editing-->
-    <string name="model_editor_delete">Delete field</string>
-    <string name="model_editor_options">Field options</string>
-    <string name="model_editor_rename">Rename field</string>
-    <string name="model_editor_reposition_menu">Reposition field</string>
-    <string name="model_editor_reposition" formatted="false">Reposition field (enter a value %1$d - %2$d)</string>
-    <string name="model_editor_changing">Updating fields</string>
-    <string name="model_editor_sort_field">Sort by this field</string>
 
+    <!--Field Editing-->
+    <string name="model_field_editor_title">Edit fields:  %1$s</string>
+    <string name="model_field_editor_delete">Delete field</string>
+    <string name="model_field_editor_options">Field options</string>
+    <string name="model_field_editor_rename">Rename field</string>
+    <string name="model_field_editor_reposition_menu">Reposition field</string>
+    <string name="model_field_editor_reposition" formatted="false">Reposition field (enter a value %1$d - %2$d)</string>
+    <string name="model_field_editor_changing">Updating fields</string>
+    <string name="model_field_editor_sort_field">Sort by this field</string>
     <string name="model_clone_suffix">copy</string>
 
+
     <!--Warnings-->
-    <string name="model_editor_delete_warning">Are you sure that you wish to delete this field?  The data contained in the field will be deleted and synchronizing with AnkiWeb servers will require a full sync.</string>
+    <string name="model_field_editor_delete_warning">Are you sure that you wish to delete this field?  The data contained in the field will be deleted and synchronizing with AnkiWeb servers will require a full sync.</string>
+
 
     <!--Progress Dialogues-->
-    <string name="model_editor_deletion_in_progress">Deleting Note Type</string>
-    <string name="model_editor_loading_models">Loading Note Types</string>
+    <string name="model_browser_deletion_in_progress">Deleting Note Type</string>
+    <string name="model_browser_loading_models">Loading Note Types</string>
+
 
     <!--Model Add Suffixes-->
     <string name="model_browser_add_add">Add: %1$s </string> <!--This prefix is used for standard note types-->
